fix (dev): rename directory namex
