+++ point-system/src/main/java/pointsystem/converter/CompanyConverter.java
@@ -1,4 +1,4 @@
-package pointsystem.converter_temp;
+package pointsystem.converter;
 
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeMap;
