+++ point-system/src/main/java/pointsystem/converter/UserConverter.java
@@ -1,4 +1,4 @@
-package pointsystem.converter_temp;
+package pointsystem.converter;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
